plan: a grid column the schema forbids was undefined behaviour
`GridPosition` is x in [-4096, 4096] and y in [-256, 256]
(packages/schemas/src/refs.ts), and the plugin enforced neither. `readIntField`
guaranteed only that the value fit a C++ `int`, and `layout::gridToPixel` then
computed `(gx + 2000) * width` in `int`. For any column above INT_MAX - 2000
that is signed overflow. Reproduced under UBSan, and the wrapped result is not
merely wrong but backwards: `x: 2147483647` placed the module at -3.2e10, far
to the LEFT of a rack whose right edge was being asked for.

The MCP server's Zod schemas refuse this first, so nothing the shipped server
sends is affected -- but docs/security/threat-model.md draws boundary 2 at the
bridge socket and promises the plugin re-validates every frame, and anything
else holding the writer lease reaches `move_module` directly. This is the same
habit the last eight commits have been closing: an accessor that accepts what
it does not understand.

Three changes:

- `scripts/gen-cpp.ts` harvests the domain from every `position` property in
  patch-operation.schema.json and emits `gen::GRID_POSITION_{X,Y}_{MIN,MAX}`.
  It refuses to generate when an occurrence carries no bounds, and when two
  occurrences disagree -- one pair of constants cannot describe two domains.
  Both guards were seen red.
- `readIntField` gains a bounded overload, and `readGridPosition` reads both
  axes. A helper rather than two calls at each of the three sites, because the
  axes do NOT share a domain: pairing the wrong constants would accept every y
  from 257 to 4096 while looking exactly like a bounds check. That mistake is
  the one the new test makes fail.
- `gridToPixel` widens before it adds. Callers now hold the schema's domain,
  but this is the function that does the arithmetic and it should not depend on
  being called correctly.

Preview refuses first, with BAD_REQUEST naming the axis and the range, so an
out-of-domain position never reaches a mutation.

+++ plugins/RackMCP/src/core/layout.cpp
@@ -28,7 +28,14 @@ bool intersects(const Box& a, const Box& b) {
 }
 
 Point gridToPixel(int gx, int gy, const Grid& grid) {
-    return Point((gx + grid.originColumns) * grid.width, (gy + grid.originRows) * grid.height);
+    // Widened before the addition, not after. Rack's origin is 2000 columns and
+    // 100 rows, so `gx + originColumns` in `int` overflows for any column above
+    // INT_MAX - 2000 -- undefined behaviour, reachable from the wire because
+    // readIntField only guaranteed the value fit an `int`. Callers now hold the
+    // schema's much narrower GridPosition domain, but this is the function that
+    // does the arithmetic and it should not depend on being called correctly.
+    return Point((float) ((int64_t) gx + grid.originColumns) * grid.width,
+                 (float) ((int64_t) gy + grid.originRows) * grid.height);
 }
 
 float rightmostEdge(const std::vector<Occupant>& occupants, const Grid& grid) {
